requirements.txt
cryptography==3.4.8
snowflake_sqlalchemy==1.3.3
